requirements.txt
google-generativeai
requests
beautifulsoup4
jinja2
